Contextual menu
- Add contextual menu for notes.
- Add notes deletetion.
